feat: made Card Container interactive & updated cursor
